voice_classifier/requirements.txt


PyTorch
NumPy
Librosa
Matplotlib









